+++ back-end/src/main/java/com/exadel/team3/backend/services/email/email_sender/Sender.java
@@ -1,7 +1,7 @@
 package com.exadel.team3.backend.services.email.email_sender;
 
 
-import com.exadel.team3.backend.services.email.email_command_type.EmailTypes;
+import com.exadel.team3.backend.services.email.email_types.EmailTypes;
 
 import java.util.Map;
